Modify : file structure refactor
Set folders for each component
