+++ src/components/index.js
@@ -1,10 +1,10 @@
-import Header from './Header';
-import Authentication from './Authentication';
-import HomeContent from './HomeContent';
-import Letter from './Letter';
-import Contact from './Contact';
+import Header from './Header/Header';
+import Authentication from './Authentication/Authentication';
+import HomeContent from './Home/HomeContent';
+import Letter from './Letter/Letter';
+import Contact from './Contact/Contact';
 import ProfileCard from './Profile/ProfileCard';
-import Logout from './Logout';
+import Logout from './Authentication/Logout';
 
 export { Header, Authentication, HomeContent,
   Letter, Contact,
