+++ src/containers/index.js
@@ -1,10 +1,10 @@
 import App from './App';
-import Login from './Login';
-import Home from './Home';
-import Signup from './Signup';
-import DropList from './DropList';
-import LetterList from './LetterList';
-import ContactList from './ContactList';
+import Login from './Authentication/Login';
+import Home from './Home/Home';
+import Signup from './Authentication/Signup';
+import DropList from './Drop/DropList';
+import LetterList from './Letter/LetterList';
+import ContactList from './Contact/ContactList';
 import Profile from './Profile/Profile';
 import List from './List/List';
